back-end-python-django/requirements.txt
Django==4.2.0
djangorestframework==3.15.1
drf-yasg
django-rest-swagger==2.2.0
mysqlclient
drf-spectacular
mysql-client